refactor: Square 이름을 Position으로 변경
